refactor(server): active or suspendent extends generic runner
